% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ballfield.R
\name{geom_ballfield}
\alias{geom_ballfield}
\title{Draw the boundaries of the infield and outfield areas}
\usage{
geom_ballfield()
}
\value{

}
\description{
The actual values for the coordinates of 1B, 2B, and 3B were eyeballed by
displaying the panel.grid and axis.text arguments as normal elements (they
are set to \code{element_blank()}) for the "real plots.
}
