% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspray_ecm.R
\name{ggspraychart_ecm}
\alias{ggspraychart_ecm}
\title{Plot the density of batted ball coordinates}
\usage{
ggspraychart_ecm(
  data,
  x_value = "hc_x",
  y_value = "-hc_y",
  low_color = "#b1d3e6",
  high_color = "#8b0000",
  fill_value = NULL,
  fill_palette = NULL,
  fill_legend_title = NULL,
  bin_size = 15,
  alpha = 0.6,
  frame = NULL
)
}
\arguments{
\item{data}{A data frame containing "hc_x" and "hc_y" from StatCast database}

\item{x_value}{Coordinates of the ball impact or fielding location}

\item{y_value}{Coordinates of the ball impact or fielding location}

\item{low_color}{Hex code for the color representing lowest density}

\item{high_color}{Hex code for the color representing highest density}

\item{fill_value}{see \code{ggspraychart()} in \strong{baseballr}}

\item{fill_palette}{A means to adjust the fill palette}

\item{fill_legend_title}{A custom legend title}

\item{bin_size}{Number of bins for computing density contours, defaults to 15}

\item{alpha}{The opacity of the filled polygons}

\item{frame}{used to generate animated ggplots}
}
\value{
a ggplot object
}
\description{
This is an adaptation of \code{ggspraychart} from the excellent \strong{baseballr}
package. The main difference is the visual thickness of the contour lines and
the color palette. Points are not used for this plot, which substantially
reduces the computational requirements for large datasets.
}
