% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspray_ecm_skeleton.R
\name{ggspraychart_ecm_skeleton}
\alias{ggspraychart_ecm_skeleton}
\title{Draw an empty spray chart on which to plot the ballfield boundaries}
\usage{
ggspraychart_ecm_skeleton(
  data,
  x_value = "hc_x",
  y_value = "-hc_y",
  low_color = "#black",
  high_color = "white",
  fill_value = NULL,
  fill_palette = NULL,
  fill_legend_title = NULL,
  bin_size = 15,
  frame = NULL
)
}
\arguments{
\item{data}{A data frame containing "hc_x" and "hc_y" from StatCast database}

\item{x_value}{Coordinates of the ball impact or fielding location}

\item{y_value}{Coordinates of the ball impact or fielding location}

\item{low_color}{Hex code for the color representing lowest density}

\item{high_color}{Hex code for the color representing highest density}

\item{fill_value}{see \code{ggspraychart()} in \strong{baseballr}}

\item{fill_palette}{A means to adjust the fill palette}

\item{fill_legend_title}{A custom legend title}

\item{bin_size}{Number of bins for computing density contours, defaults to 15}

\item{frame}{used to generate animated ggplots}
}
\value{

}
\description{
A wrapper around ggspraychart_ecm that allows no data to be plotted, leaving the plot area
empty and able to accept a small dummy data set for adding a \code{geom_ballfield()}. Use to draw a
new plot with only one layer of \code{geom_ballfield()} instead of the hugs file that results from
using normal data sets.
}
