% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ballfield.R
\name{theme_ballfield}
\alias{theme_ballfield}
\title{Adjust theme elements of a a \code{ggspray_ecm} plot \strong{without} altering any font sizes (this will be left to \strong{cowplot})}
\usage{
theme_ballfield(base_size = 12)
}
\arguments{
\item{base_size}{font size}
}
\value{

}
\description{
This theme removes undesired components and sets the parameters for the legend, # etc.
}
