% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ballfield_fixed_font_sizes.R
\name{theme_ballfield_fixed_font_sizes}
\alias{theme_ballfield_fixed_font_sizes}
\title{Adjust theme elements of a a \code{ggspray_ecm} plot}
\usage{
theme_ballfield_fixed_font_sizes(base_size = 12, base_family = "Arial Narrow")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\value{

}
\description{
This theme removes undesired components and sets the parameters for the legend, # etc.
}
