% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ballfield_fonts_untouched.R
\name{theme_ballfield_fonts_untouched}
\alias{theme_ballfield_fonts_untouched}
\title{Adjust theme elements of a a \code{ggspray_ecm} plot \strong{without} altering any font sizes (this will be left to \strong{cowplot})}
\usage{
theme_ballfield_fonts_untouched(base_size = 12, base_family = "Arial Narrow")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\value{

}
\description{
This theme removes undesired components and sets the parameters for the legend, # etc.
}
